% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploGeno-package.R
\docType{package}
\name{HaploGeno-package}
\alias{HaploGeno}
\alias{HaploGeno-package}
\title{HaploGeno: Haplotype-Based Genomic Prediction}
\description{
Implements a high-performance framework for haplotype-based genomic prediction and QTL discovery. Leverages file-backed matrices via 'bigstatsr' to handle large-scale genomic datasets efficiently. Key features include LD-based haplotype block construction, kernel ridge regression (KRR) for genomic prediction, and local genomic estimated breeding value (GEBV) analysis for identifying significant genomic regions.
}
\author{
\strong{Maintainer}: Zachary Aldiss \email{zaldiss@illinois.edu}

}
\keyword{internal}
