% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploObject.R
\name{load_haplo_project}
\alias{load_haplo_project}
\title{Load HaploGeno Project}
\usage{
load_haplo_project(path)
}
\arguments{
\item{path}{Path to the .rds file.}
}
\value{
A HaploObject instance.
}
\description{
Load HaploGeno Project
}
