% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Methods.R
\name{plot.HaploObject}
\alias{plot.HaploObject}
\title{Plot method for HaploObject}
\usage{
\method{plot}{HaploObject}(x, type = "manhattan", ...)
}
\arguments{
\item{x}{A HaploObject}

\item{type}{Character string. "manhattan" (default).}

\item{...}{Additional arguments passed to plotting methods.}
}
\description{
Plot method for HaploObject
}
