% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Methods.R
\name{print.HaploObject}
\alias{print.HaploObject}
\title{Print method for HaploObject}
\usage{
\method{print}{HaploObject}(x, ...)
}
\arguments{
\item{x}{A HaploObject}

\item{...}{Additional arguments}
}
\description{
Print method for HaploObject
}
