% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTrait.R
\name{print.MultiTraitFA}
\alias{print.MultiTraitFA}
\title{Print Multi-Trait FA Results}
\usage{
\method{print}{MultiTraitFA}(x, ...)
}
\arguments{
\item{x}{A MultiTraitFA object.}

\item{...}{Additional arguments.}
}
\description{
Print Multi-Trait FA Results
}
