% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoi_analysis.R
\name{scale_haplo_effects}
\alias{scale_haplo_effects}
\title{Scale Haplotype Effects}
\usage{
scale_haplo_effects(local_gebv)
}
\arguments{
\item{local_gebv}{A matrix of local GEBVs (N x n_blocks).}
}
\value{
A matrix of scaled local GEBVs.
}
\description{
Centers the local GEBV matrix columns to have a mean of zero.
}
