% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Methods.R
\name{summary.HaploObject}
\alias{summary.HaploObject}
\title{Summary method for HaploObject}
\usage{
\method{summary}{HaploObject}(object, ...)
}
\arguments{
\item{object}{A HaploObject}

\item{...}{Additional arguments}
}
\description{
Summary method for HaploObject
}
