% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploObject.R
\name{HaploObject}
\alias{HaploObject}
\title{HaploObject Class}
\description{
HaploObject Class

HaploObject Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{geno}}{Filebacked Big Matrix of genotypes.}

\item{\code{map}}{Data.table containing marker map information.}

\item{\code{pheno}}{Vector of phenotypes.}

\item{\code{blocks}}{Data.table defining haplotype blocks.}

\item{\code{haplo_mat}}{Matrix of encoded haplotypes.}

\item{\code{hrm}}{Haplotype Relationship Matrix.}

\item{\code{marker_effects}}{Vector of estimated marker effects.}

\item{\code{local_gebv}}{List containing local GEBV matrix and variances.}

\item{\code{significance}}{Data.table of significance test results.}

\item{\code{active_markers}}{Integer vector of indices of markers with non-zero variance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HaploObject-new}{\code{HaploObject$new()}}
\item \href{#method-HaploObject-print}{\code{HaploObject$print()}}
\item \href{#method-HaploObject-import_genotypes}{\code{HaploObject$import_genotypes()}}
\item \href{#method-HaploObject-load_map}{\code{HaploObject$load_map()}}
\item \href{#method-HaploObject-load_pheno}{\code{HaploObject$load_pheno()}}
\item \href{#method-HaploObject-get_subset}{\code{HaploObject$get_subset()}}
\item \href{#method-HaploObject-define_blocks_fixed}{\code{HaploObject$define_blocks_fixed()}}
\item \href{#method-HaploObject-define_blocks_ld}{\code{HaploObject$define_blocks_ld()}}
\item \href{#method-HaploObject-estimate_marker_effects}{\code{HaploObject$estimate_marker_effects()}}
\item \href{#method-HaploObject-encode_haplotypes}{\code{HaploObject$encode_haplotypes()}}
\item \href{#method-HaploObject-calculate_local_gebv}{\code{HaploObject$calculate_local_gebv()}}
\item \href{#method-HaploObject-test_significance}{\code{HaploObject$test_significance()}}
\item \href{#method-HaploObject-compute_hrm}{\code{HaploObject$compute_hrm()}}
\item \href{#method-HaploObject-fit_krr}{\code{HaploObject$fit_krr()}}
\item \href{#method-HaploObject-cross_validate}{\code{HaploObject$cross_validate()}}
\item \href{#method-HaploObject-plot_manhattan}{\code{HaploObject$plot_manhattan()}}
\item \href{#method-HaploObject-plot_pca}{\code{HaploObject$plot_pca()}}
\item \href{#method-HaploObject-plot_gebv_image}{\code{HaploObject$plot_gebv_image()}}
\item \href{#method-HaploObject-identify_superior_haplotypes}{\code{HaploObject$identify_superior_haplotypes()}}
\item \href{#method-HaploObject-score_stacking}{\code{HaploObject$score_stacking()}}
\item \href{#method-HaploObject-plot_stacking_trend}{\code{HaploObject$plot_stacking_trend()}}
\item \href{#method-HaploObject-impute_genotypes}{\code{HaploObject$impute_genotypes()}}
\item \href{#method-HaploObject-filter_monomorphic}{\code{HaploObject$filter_monomorphic()}}
\item \href{#method-HaploObject-save_project}{\code{HaploObject$save_project()}}
\item \href{#method-HaploObject-clone}{\code{HaploObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new HaploObject.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$new(backing_file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backing_file_path}}{Path to the backing file for the FBM.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the HaploObject.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-import_genotypes"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-import_genotypes}{}}}
\subsection{Method \code{import_genotypes()}}{
Import genotypes from a file or matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$import_genotypes(matrix_or_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_or_path}}{A matrix or path to a file (bed/vcf/rds).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-load_map"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-load_map}{}}}
\subsection{Method \code{load_map()}}{
Load marker map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$load_map(map_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map_data}}{A data.frame or path to a file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-load_pheno"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-load_pheno}{}}}
\subsection{Method \code{load_pheno()}}{
Load phenotypes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$load_pheno(pheno_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pheno_data}}{A vector, data.frame, or path to a file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-get_subset"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-get_subset}{}}}
\subsection{Method \code{get_subset()}}{
Get a subset of the genotype matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$get_subset(row_ind, col_ind)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row_ind}}{Row indices.}

\item{\code{col_ind}}{Column indices.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-define_blocks_fixed"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-define_blocks_fixed}{}}}
\subsection{Method \code{define_blocks_fixed()}}{
Define haplotype blocks using fixed window size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$define_blocks_fixed(window_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{window_size}}{Number of markers per block.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-define_blocks_ld"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-define_blocks_ld}{}}}
\subsection{Method \code{define_blocks_ld()}}{
Define haplotype blocks using LD scan.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$define_blocks_ld(
  r2_threshold = 0.1,
  tolerance = 3,
  window_size = 2000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r2_threshold}}{r2 threshold for block definition.}

\item{\code{tolerance}}{Number of failures allowed before ending a block.}

\item{\code{window_size}}{Maximum window size to scan.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-estimate_marker_effects"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-estimate_marker_effects}{}}}
\subsection{Method \code{estimate_marker_effects()}}{
Estimate marker effects using Ridge Regression.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$estimate_marker_effects(lambda = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{Regularization parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-encode_haplotypes"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-encode_haplotypes}{}}}
\subsection{Method \code{encode_haplotypes()}}{
Encode haplotypes into integer IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$encode_haplotypes(n_cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_cores}}{Number of cores to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-calculate_local_gebv"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-calculate_local_gebv}{}}}
\subsection{Method \code{calculate_local_gebv()}}{
Calculate local GEBVs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$calculate_local_gebv(n_cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_cores}}{Number of cores to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-test_significance"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-test_significance}{}}}
\subsection{Method \code{test_significance()}}{
Test significance of local GEBVs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$test_significance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data.table of p-values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-compute_hrm"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-compute_hrm}{}}}
\subsection{Method \code{compute_hrm()}}{
Compute Haplotype Relationship Matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$compute_hrm(n_cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_cores}}{Number of cores to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-fit_krr"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-fit_krr}{}}}
\subsection{Method \code{fit_krr()}}{
Fit Kernel Ridge Regression model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$fit_krr(lambda = 0.1, use_cg = NULL, tol = 1e-05, max_iter = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{Regularization parameter.}

\item{\code{use_cg}}{Whether to use Conjugate Gradient solver.}

\item{\code{tol}}{Tolerance for CG.}

\item{\code{max_iter}}{Maximum iterations for CG.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-cross_validate"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-cross_validate}{}}}
\subsection{Method \code{cross_validate()}}{
Cross-validate KRR model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$cross_validate(k = 5, lambdas = NULL, n_cores = 1, folds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{Number of folds.}

\item{\code{lambdas}}{Vector of lambdas to test.}

\item{\code{n_cores}}{Number of cores to use.}

\item{\code{folds}}{Optional vector of fold assignments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-plot_manhattan"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-plot_manhattan}{}}}
\subsection{Method \code{plot_manhattan()}}{
Plot Manhattan plot of local GEBV significance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$plot_manhattan(threshold = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{Significance threshold (p-value).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-plot_pca"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-plot_pca}{}}}
\subsection{Method \code{plot_pca()}}{
Plot PCA of HRM.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$plot_pca(groups = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{Optional vector of groups for coloring.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-plot_gebv_image"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-plot_gebv_image}{}}}
\subsection{Method \code{plot_gebv_image()}}{
Plot heatmap of local GEBVs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$plot_gebv_image(block_range = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block_range}}{Optional range of blocks to plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-identify_superior_haplotypes"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-identify_superior_haplotypes}{}}}
\subsection{Method \code{identify_superior_haplotypes()}}{
Identify superior haplotypes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$identify_superior_haplotypes(top_n = 50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{top_n}}{Number of top blocks to consider.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-score_stacking"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-score_stacking}{}}}
\subsection{Method \code{score_stacking()}}{
Calculate stacking scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$score_stacking(superior_haplos)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{superior_haplos}}{Data.table of superior haplotypes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-plot_stacking_trend"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-plot_stacking_trend}{}}}
\subsection{Method \code{plot_stacking_trend()}}{
Plot stacking trend.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$plot_stacking_trend(scores = NULL, superior_haplos = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scores}}{Vector of stacking scores.}

\item{\code{superior_haplos}}{Optional table of superior haplotypes (if scores not provided).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-impute_genotypes"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-impute_genotypes}{}}}
\subsection{Method \code{impute_genotypes()}}{
Impute missing genotypes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$impute_genotypes(method = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Imputation method. Currently only "mean" is supported.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-filter_monomorphic"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-filter_monomorphic}{}}}
\subsection{Method \code{filter_monomorphic()}}{
Filter monomorphic markers (zero variance).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$filter_monomorphic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-save_project"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-save_project}{}}}
\subsection{Method \code{save_project()}}{
Save the project to an RDS file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$save_project(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the output .rds file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HaploObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HaploObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HaploObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
